+++ backend/tests/fixtures/table.py
@@ -1,4 +1,5 @@
-from django.db import connection
+from django.conf import settings
+from django.db import connections
 
 from baserow.contrib.database.table.models import Table
 
@@ -17,7 +18,8 @@ class TableFixtures:
         table = Table.objects.create(**kwargs)
 
         if create_table:
-            with connection.schema_editor() as schema_editor:
+            user_table_db = connections[settings.USER_TABLE_DATABASE]
+            with user_table_db.schema_editor() as schema_editor:
                 schema_editor.create_model(table.get_model())
 
         return table