+++ changelog.md
@@ -20,6 +20,7 @@
 * Fixed 100X backend web socket errors when refreshing the page.
 * Fixed SSRF bug in the file upload by URL by blocking urls to the private network.
 * Fixed bug where an invalid date could be converted to 0001-01-01.
+* The list_database_table_rows search query parameter now searches all possible field types.
 
 ## Released (2021-03-01)